Simplify
previous modus operandi was:

- copy terraform configs (generated or otherwise) into a store path
- do some very hacky gymnastics to get terraform to run from that store
path while writing all state to the original directory it was run from

this works, but:
- it's messy and hard to reason about
- "regular" terraform modules can't access sources using relative paths

Instead, we simply generate a single `nix.tf.json` store path from all `tf.nix` files in the current directory and link to it. So at the cost of having to `.gitignore` a file, we get to eliminate a bunch of messy code and support more use cases.